+++ lib/internal/bootstrap/node.js
@@ -52,7 +52,7 @@ const {
   globalThis,
 } = primordials;
 const config = internalBinding('config');
-const { deprecate, lazyDOMExceptionClass } = require('internal/util');
+const { deprecate } = require('internal/util');
 
 setupProcessObject();
 
@@ -184,82 +184,7 @@ if (credentials.implementsPosixCredentials) {
 const { nativeHooks } = require('internal/async_hooks');
 internalBinding('async_wrap').setupHooks(nativeHooks);
 
-const {
-  setupTaskQueue,
-  queueMicrotask
-} = require('internal/process/task_queues');
-
-if (!config.noBrowserGlobals) {
-  // Override global console from the one provided by the VM
-  // to the one implemented by Node.js
-  // https://console.spec.whatwg.org/#console-namespace
-  exposeNamespace(globalThis, 'console',
-                  createGlobalConsole(globalThis.console));
-
-  const { URL, URLSearchParams } = require('internal/url');
-  // https://url.spec.whatwg.org/#url
-  exposeInterface(globalThis, 'URL', URL);
-  // https://url.spec.whatwg.org/#urlsearchparams
-  exposeInterface(globalThis, 'URLSearchParams', URLSearchParams);
-  exposeGetterAndSetter(globalThis,
-                        'DOMException',
-                        lazyDOMExceptionClass,
-                        (value) => {
-                          exposeInterface(globalThis, 'DOMException', value);
-                        });
-
-  const {
-    TextEncoder, TextDecoder
-  } = require('internal/encoding');
-  // https://encoding.spec.whatwg.org/#textencoder
-  exposeInterface(globalThis, 'TextEncoder', TextEncoder);
-  // https://encoding.spec.whatwg.org/#textdecoder
-  exposeInterface(globalThis, 'TextDecoder', TextDecoder);
-
-  const {
-    AbortController,
-    AbortSignal,
-  } = require('internal/abort_controller');
-  exposeInterface(globalThis, 'AbortController', AbortController);
-  exposeInterface(globalThis, 'AbortSignal', AbortSignal);
-
-  const {
-    EventTarget,
-    Event,
-  } = require('internal/event_target');
-  exposeInterface(globalThis, 'EventTarget', EventTarget);
-  exposeInterface(globalThis, 'Event', Event);
-  const {
-    MessageChannel,
-    MessagePort,
-    MessageEvent,
-  } = require('internal/worker/io');
-  exposeInterface(globalThis, 'MessageChannel', MessageChannel);
-  exposeInterface(globalThis, 'MessagePort', MessagePort);
-  exposeInterface(globalThis, 'MessageEvent', MessageEvent);
-
-  // https://html.spec.whatwg.org/multipage/webappapis.html#windoworworkerglobalscope
-  const timers = require('timers');
-  defineOperation(globalThis, 'clearInterval', timers.clearInterval);
-  defineOperation(globalThis, 'clearTimeout', timers.clearTimeout);
-  defineOperation(globalThis, 'setInterval', timers.setInterval);
-  defineOperation(globalThis, 'setTimeout', timers.setTimeout);
-
-  defineOperation(globalThis, 'queueMicrotask', queueMicrotask);
-
-  // https://www.w3.org/TR/hr-time-2/#the-performance-attribute
-  defineReplacableAttribute(globalThis, 'performance',
-                            require('perf_hooks').performance);
-
-  // Non-standard extensions:
-  defineOperation(globalThis, 'clearImmediate', timers.clearImmediate);
-  defineOperation(globalThis, 'setImmediate', timers.setImmediate);
-
-  const {
-    structuredClone,
-  } = require('internal/structured_clone');
-  defineOperation(globalThis, 'structuredClone', structuredClone);
-}
+const { setupTaskQueue } = require('internal/process/task_queues');
 
 // Set the per-Environment callback that will be called
 // when the TrackingTraceStateObserver updates trace state.
@@ -454,71 +379,5 @@ function setupBuffer() {
       writable: true,
       configurable: true,
     },
-  });
-}
-
-function createGlobalConsole(consoleFromVM) {
-  const consoleFromNode =
-    require('internal/console/global');
-  if (config.hasInspector) {
-    const inspector = require('internal/util/inspector');
-    // This will be exposed by `require('inspector').console` later.
-    inspector.consoleFromVM = consoleFromVM;
-    // TODO(joyeecheung): postpone this until the first time inspector
-    // is activated.
-    inspector.wrapConsole(consoleFromNode, consoleFromVM);
-    const { setConsoleExtensionInstaller } = internalBinding('inspector');
-    // Setup inspector command line API.
-    setConsoleExtensionInstaller(inspector.installConsoleExtensions);
-  }
-  return consoleFromNode;
-}
-
-// https://heycam.github.io/webidl/#es-namespaces
-function exposeNamespace(target, name, namespaceObject) {
-  ObjectDefineProperty(target, name, {
-    writable: true,
-    enumerable: false,
-    configurable: true,
-    value: namespaceObject
-  });
-}
-
-// https://heycam.github.io/webidl/#es-interfaces
-function exposeInterface(target, name, interfaceObject) {
-  ObjectDefineProperty(target, name, {
-    writable: true,
-    enumerable: false,
-    configurable: true,
-    value: interfaceObject
-  });
-}
-
-function exposeGetterAndSetter(target, name, getter, setter = undefined) {
-  ObjectDefineProperty(target, name, {
-    enumerable: false,
-    configurable: true,
-    get: getter,
-    set: setter,
-  });
-}
-
-// https://heycam.github.io/webidl/#define-the-operations
-function defineOperation(target, name, method) {
-  ObjectDefineProperty(target, name, {
-    writable: true,
-    enumerable: true,
-    configurable: true,
-    value: method
-  });
-}
-
-// https://heycam.github.io/webidl/#Replaceable
-function defineReplacableAttribute(target, name, value) {
-  ObjectDefineProperty(target, name, {
-    writable: true,
-    enumerable: true,
-    configurable: true,
-    value,
   });
 }
