+++ lib/internal/bootstrap/pre_execution.js
@@ -17,12 +17,17 @@ const { reconnectZeroFillToggle } = require('internal/buffer');
 
 const { Buffer } = require('buffer');
 const { ERR_MANIFEST_ASSERT_INTEGRITY } = require('internal/errors').codes;
+const { lazyDOMExceptionClass } = require('internal/util');
 const assert = require('internal/assert');
+const config = internalBinding('config');
 
 function prepareMainThreadExecution(expandArgv1 = false) {
   // TODO(joyeecheung): this is also necessary for workers when they deserialize
   // this toggle from the snapshot.
   reconnectZeroFillToggle();
+
+  // Create globals exposing browser interfaces.
+  createBrowserGlobals();
 
   // Patch the process object with legacy properties and normalizations
   patchProcessObject(expandArgv1);
@@ -76,6 +81,148 @@ function prepareMainThreadExecution(expandArgv1 = false) {
   assert(!CJSLoader.hasLoadedAnyUserCJSModule);
   loadPreloadModules();
   initializeFrozenIntrinsics();
+}
+
+function createBrowserGlobals() {
+  if (getEmbedderOptions().noBrowserGlobals)
+    return;
+
+  // Override global console from the one provided by the VM
+  // to the one implemented by Node.js
+  // https://console.spec.whatwg.org/#console-namespace
+  exposeNamespace(globalThis, 'console',
+                  createGlobalConsole(globalThis.console));
+
+  const { URL, URLSearchParams } = require('internal/url');
+  // https://url.spec.whatwg.org/#url
+  exposeInterface(globalThis, 'URL', URL);
+  // https://url.spec.whatwg.org/#urlsearchparams
+  exposeInterface(globalThis, 'URLSearchParams', URLSearchParams);
+  exposeGetterAndSetter(globalThis,
+                        'DOMException',
+                        lazyDOMExceptionClass,
+                        (value) => {
+                          exposeInterface(globalThis, 'DOMException', value);
+                        });
+
+  const {
+    TextEncoder, TextDecoder
+  } = require('internal/encoding');
+  // https://encoding.spec.whatwg.org/#textencoder
+  exposeInterface(globalThis, 'TextEncoder', TextEncoder);
+  // https://encoding.spec.whatwg.org/#textdecoder
+  exposeInterface(globalThis, 'TextDecoder', TextDecoder);
+
+  const {
+    AbortController,
+    AbortSignal,
+  } = require('internal/abort_controller');
+  exposeInterface(globalThis, 'AbortController', AbortController);
+  exposeInterface(globalThis, 'AbortSignal', AbortSignal);
+
+  const {
+    EventTarget,
+    Event,
+  } = require('internal/event_target');
+  exposeInterface(globalThis, 'EventTarget', EventTarget);
+  exposeInterface(globalThis, 'Event', Event);
+  const {
+    MessageChannel,
+    MessagePort,
+    MessageEvent,
+  } = require('internal/worker/io');
+  exposeInterface(globalThis, 'MessageChannel', MessageChannel);
+  exposeInterface(globalThis, 'MessagePort', MessagePort);
+  exposeInterface(globalThis, 'MessageEvent', MessageEvent);
+
+  // https://html.spec.whatwg.org/multipage/webappapis.html#windoworworkerglobalscope
+  const timers = require('timers');
+  defineOperation(globalThis, 'clearInterval', timers.clearInterval);
+  defineOperation(globalThis, 'clearTimeout', timers.clearTimeout);
+  defineOperation(globalThis, 'setInterval', timers.setInterval);
+  defineOperation(globalThis, 'setTimeout', timers.setTimeout);
+
+  const { queueMicrotask } = require('internal/process/task_queues');
+  defineOperation(globalThis, 'queueMicrotask', queueMicrotask);
+
+  // https://www.w3.org/TR/hr-time-2/#the-performance-attribute
+  defineReplacableAttribute(globalThis, 'performance',
+                            require('perf_hooks').performance);
+
+  // Non-standard extensions:
+  defineOperation(globalThis, 'clearImmediate', timers.clearImmediate);
+  defineOperation(globalThis, 'setImmediate', timers.setImmediate);
+
+  const {
+    structuredClone,
+  } = require('internal/structured_clone');
+  defineOperation(globalThis, 'structuredClone', structuredClone);
+}
+
+function createGlobalConsole(consoleFromVM) {
+  const consoleFromNode =
+    require('internal/console/global');
+  if (config.hasInspector) {
+    const inspector = require('internal/util/inspector');
+    // This will be exposed by `require('inspector').console` later.
+    inspector.consoleFromVM = consoleFromVM;
+    // TODO(joyeecheung): postpone this until the first time inspector
+    // is activated.
+    inspector.wrapConsole(consoleFromNode, consoleFromVM);
+    const { setConsoleExtensionInstaller } = internalBinding('inspector');
+    // Setup inspector command line API.
+    setConsoleExtensionInstaller(inspector.installConsoleExtensions);
+  }
+  return consoleFromNode;
+}
+
+// https://heycam.github.io/webidl/#es-namespaces
+function exposeNamespace(target, name, namespaceObject) {
+  ObjectDefineProperty(target, name, {
+    writable: true,
+    enumerable: false,
+    configurable: true,
+    value: namespaceObject
+  });
+}
+
+// https://heycam.github.io/webidl/#es-interfaces
+function exposeInterface(target, name, interfaceObject) {
+  ObjectDefineProperty(target, name, {
+    writable: true,
+    enumerable: false,
+    configurable: true,
+    value: interfaceObject
+  });
+}
+
+function exposeGetterAndSetter(target, name, getter, setter = undefined) {
+  ObjectDefineProperty(target, name, {
+    enumerable: false,
+    configurable: true,
+    get: getter,
+    set: setter,
+  });
+}
+
+// https://heycam.github.io/webidl/#define-the-operations
+function defineOperation(target, name, method) {
+  ObjectDefineProperty(target, name, {
+    writable: true,
+    enumerable: true,
+    configurable: true,
+    value: method
+  });
+}
+
+// https://heycam.github.io/webidl/#Replaceable
+function defineReplacableAttribute(target, name, value) {
+  ObjectDefineProperty(target, name, {
+    writable: true,
+    enumerable: true,
+    configurable: true,
+    value,
+  });
 }
 
 function patchProcessObject(expandArgv1) {
@@ -286,7 +433,7 @@ function initializeDeprecations() {
   // Now that we use the config binding to carry this information, remove
   // it from the process. We may consider exposing it properly in
   // process.features.
-  const { noBrowserGlobals } = internalBinding('config');
+  const { noBrowserGlobals } = getEmbedderOptions();
   if (noBrowserGlobals) {
     ObjectDefineProperty(process, '_noBrowserGlobals', {
       writable: false,
@@ -485,6 +632,7 @@ function loadPreloadModules() {
 }
 
 module.exports = {
+  createBrowserGlobals,
   patchProcessObject,
   setupCoverageHooks,
   setupWarningHandler,
