+++ lib/internal/main/worker_thread.js
@@ -13,6 +13,7 @@ const {
 } = primordials;
 
 const {
+  createBrowserGlobals,
   patchProcessObject,
   setupCoverageHooks,
   setupInspectorHooks,
@@ -62,6 +63,7 @@ let debug = require('internal/util/debuglog').debuglog('worker', (fn) => {
 
 const assert = require('internal/assert');
 
+createBrowserGlobals();
 patchProcessObject();
 setupInspectorHooks();
 setupDebugEnv();
